Create ValidatePrice Create SelectPrice class for random use Price Use only BigDecimal Some update

+++ price-calc/src/main/java/by/prohorov/calc/Calculate.java
@@ -1,10 +1,12 @@
 package by.prohorov.calc;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Artsiom Prokharau 08.02.2021
  */
 
 public interface Calculate {
 
-    double costOfDelivery();
+    BigDecimal costOfDelivery();
 }
