+++ price-calc/src/main/java/by/prohorov/price/kg/CostByPriceKg.java
@@ -1,10 +1,13 @@
 package by.prohorov.price.kg;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Artsiom Prokharau 08.02.2021
  */
 
 
 public interface CostByPriceKg {
-    double costByPriceWeightInKg();
+
+    BigDecimal costByPriceWeightInKg();
 }
